文字列があるかないかのチェックをより簡単に

+++ x264guiEx/prm/auo_util.cpp
@@ -79,7 +79,7 @@ void get_aviutl_dir(char *aviutl_dir, size_t nSize) {
 //文字列の置換に必要な領域を計算する
 size_t calc_replace_mem_required(char *str, const char *old_str, const char *new_str) {
 	size_t size = strlen(str) + 1;
-	const int move_len = strlen(new_str) - strlen(old_str);
+	const size_t move_len = strlen(new_str) - strlen(old_str);
 	if (move_len <= 0)
 		return size;
 	char *p = str;
